backend/requirements.txt
fastapi
uvicorn
google-genai
ibm-watsonx-ai
python-dotenv